+++ CHANGELOG.md
@@ -20,6 +20,7 @@ _This release is scheduled to be released on 2021-04-01._
 - Respect parameter ColoredSymbolOnly also for custom events
 - Added a new parameter to hide time portion on relative times
 - Added locale to sample config file
+- Added support for self-signed certificates for the default calendar module (#466)
 
 ### Updated